test(portcullis): split the quarantine tests out, under the line ratchet
The #18 eviction regression pushed crates/portcullis/src/flow_graph_tests.rs
from 2497 to 2548 lines, past the line ratchet's 2500 default. My own commit,
caught two ticks later.

Two reasons it took that long, both worth naming. It was green when #18 landed
because the ratchet's sweep was different before this branch rebased onto a
newer main (912 files then, 658 now). And I had been running the affected-crate
tests rather than the gate-of-gates, which is the harness that actually reports
this — a gate you do not run is a gate that is not gating you.

The fix is to move the tests, not to raise a ceiling. The line ratchet only
shrinks; giving one test file a bespoke higher ceiling to accommodate a test I
added this week is exactly the move the ratchet exists to refuse.

flow_graph_quarantine_tests.rs is included by flow_graph.rs alongside its
sibling. flow_graph_tests.rs returns to 2497, and the sweep is clean at 658
files.
